# public/api-docs/openapi/movies.yaml
openapi: 3.0.2
info:
  title: Ply Movies API
  license:
    name: MIT License
    url: https://github.com/ply-ct/ply-movies/blob/master/LICENSE
  version: 1.1.1
  x-logo:
    url: /assets/img/ply.png
    altText: Ply
externalDocs:
  description: Find out more about Ply
  url: https://ply-ct.com
servers:
  - url: https://ply-ct.com
tags:
  - name: Ply Movies
    description: Horror movies from the 1930s
    externalDocs:
      description: Ply Example API
      url: https://ply-ct.com
paths:
  /movies:
    get:
      summary: List movies
      tags:
        - Movies
      parameters:
        - name: title
          in: query
          description: Movie title
          schema:
            type: string
        - name: year
          in: query
          description: Year of release
          schema:
            type: integer
            minimum: 1930
            maximum: 1939
        - name: rating
          in: query
          description: Movie rating
          schema:
            type: number
            minimum: 0
            maximum: 5
            multipleOf: 0.5
        - name: search
          in: query
          description: Find pattern
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
        - name: descending
          in: query
          description: Sort descending
          schema:
            type: boolean
        - name: start
          in: query
          description: Pagination start index
          schema:
            type: integer
        - name: max
          in: query
          description: Page size
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      summary: Create a new movie
      tags:
        - Movies
      requestBody:
        description: Movie
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '400':
          description: Invalid movie content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  '/movies/{movieId}':
    get:
      summary: Retrieve a movie
      tags:
        - Movie
      parameters:
        - name: movieId
          in: path
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
    put:
      summary: Replace a movie
      tags:
        - Movie
      parameters:
        - name: movieId
          in: path
          required: true
      requestBody:
        description: Movie
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Invalid movie content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
    patch:
      summary: Update a movie
      tags:
        - Movie
      parameters:
        - name: movieId
          in: path
          required: true
      requestBody:
        description: Partial movie
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Invalid movie content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
    delete:
      summary: Delete a movie
      tags:
        - Movie
      parameters:
        - name: movieId
          in: path
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
components:
  schemas:
    Credit:
      additionalProperties: false
      description: Credited cast member
      properties:
        name:
          description: Cast member's name
          type: string
        role:
          description: Their role in the production
          type: enum
            - actor
            - director
      required:
        - name
        - role
      type: object
    WebRef:
      additionalProperties: false
      description: Web reference with movie details
      properties:
        site:
          description: Movie info site (imdb.com, themoviedb.org)
          type: string
        ref:
          description: The site's reference id for this movie
          type: string
      required:
        - site
        - ref
      type: object
    Movie:
      additionalProperties: false
      description: Movie
      properties:
        id:
          description: Unique ID provided by the API
          type: string
          readOnly: true
        title:
          description: Movie title
          type: string
        year:
          description: Year of release
          type: integer
          minimum: 1930
          maximum: 1939
        rating:
          description: Movie rating
          type: number
          minimum: 0
          maximum: 5
          multipleOf: 0.5
        credits:
          description: Credited cast members
          type: array
          items:
            $ref: '#/components/schemas/Credit'
        description:
          description: One-sentence discussion
        poster:
          description: Poster image file path
          type: string
          maxLength: 2048
        webRef:
          $ref: '#/components/schemas/WebRef'
      required:
        - id
        - title
        - year
      type: object
    Status:
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/StatusCodes'
        errors:
          items:
            type: string
          type: array
        message:
          type: string
      required:
        - code
        - message
      type: object
    StatusResponse:
      additionalProperties: false
      properties:
        status:
          $ref: '#/components/schemas/Status'
      required:
        - status
      type: object
    StatusCodes:
      enum:
        - 100
        - 101
        - 102
        - 200
        - 201
        - 202
        - 203
        - 204
        - 205
        - 206
        - 207
        - 300
        - 301
        - 302
        - 303
        - 304
        - 305
        - 307
        - 308
        - 400
        - 401
        - 402
        - 403
        - 404
        - 405
        - 406
        - 407
        - 408
        - 409
        - 410
        - 411
        - 412
        - 413
        - 414
        - 415
        - 416
        - 417
        - 418
        - 419
        - 420
        - 422
        - 423
        - 424
        - 428
        - 429
        - 431
        - 451
        - 500
        - 501
        - 502
        - 503
        - 504
        - 505
        - 507
        - 511
      type: number